refactor(google): move schema into separate file
